Move `NoRowsError` to `lib/rdbms`
